templates/api/django/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.8.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==24.2
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.7
gunicorn==22.0.0
whitenoise==6.7.0